kvm: Add Agent Helper to investigate if a Host is truly down
The additional webserver on port 8080 exposes a very simple JSON API
which returns the amount and list of Virtual Machines still running
on that host according to libvirt.

In case of HA the KVMHAInvestigator can perform a HTTP-call to this
simple webserver and determine if the host is actually down or if
it is just the Java Agent which has crashed.